Changed package name from ranking.advanced -> ranking.algorithms
