+++ src/main/java/edu/ucsf/rbvi/clusterMaker2/internal/CyActivator.java
@@ -30,8 +30,8 @@ import edu.ucsf.rbvi.clusterMaker2.internal.algorithms.networkClusterers.SCPS.SC
 import edu.ucsf.rbvi.clusterMaker2.internal.algorithms.networkClusterers.TransClust.TransClustClusterTaskFactory;
 import edu.ucsf.rbvi.clusterMaker2.internal.algorithms.pca.PCAMenuTaskFactory;
 import edu.ucsf.rbvi.clusterMaker2.internal.algorithms.pca.PCATaskFactory;
-import edu.ucsf.rbvi.clusterMaker2.internal.algorithms.ranking.advanced.MAATaskFactory;
-import edu.ucsf.rbvi.clusterMaker2.internal.algorithms.ranking.advanced.MAMTaskFactory;
+import edu.ucsf.rbvi.clusterMaker2.internal.algorithms.ranking.algorithms.MAATaskFactory;
+import edu.ucsf.rbvi.clusterMaker2.internal.algorithms.ranking.algorithms.MAMTaskFactory;
 import edu.ucsf.rbvi.clusterMaker2.internal.api.ClusterTaskFactory;
 import edu.ucsf.rbvi.clusterMaker2.internal.api.ClusterVizFactory;
 import edu.ucsf.rbvi.clusterMaker2.internal.api.RankFactory;
