+++ src/main/java/edu/ucsf/rbvi/clusterMaker2/internal/algorithms/ranking/algorithms/MultipleAttributeMultiplicative.java
@@ -1,4 +1,4 @@
-package edu.ucsf.rbvi.clusterMaker2.internal.algorithms.ranking.advanced;
+package edu.ucsf.rbvi.clusterMaker2.internal.algorithms.ranking.algorithms;
 
 import edu.ucsf.rbvi.clusterMaker2.internal.algorithms.NodeCluster;
 import edu.ucsf.rbvi.clusterMaker2.internal.api.ClusterManager;
@@ -13,22 +13,22 @@ import org.cytoscape.work.Tunable;
 import java.util.ArrayList;
 import java.util.List;
 
-public class MultipleAttributeAddition extends AbstractTask implements Rank {
+public class MultipleAttributeMultiplicative extends AbstractTask implements Rank {
     private List<NodeCluster> clusters;
     private ClusterManager manager;
     private List<String> nodeAttributes;
     private List<String> edgeAttributes;
     private String clusterColumnName;
-    final public static String NAME = "Create rank from multiple nodes and edges (additive sum)";
-    final public static String SHORTNAME = "MAA";
+    final public static String NAME = "Create rank from multiple nodes and edges (multiply sum)";
+    final public static String SHORTNAME = "MAM";
 
     @Tunable(description = "Network", context = "nogui")
     public CyNetwork network;
 
     @ContainsTunables
-    public MAAContext context;
+    public MAMContext context;
 
-    public MultipleAttributeAddition(MAAContext context, ClusterManager manager) {
+    public MultipleAttributeMultiplicative(MAMContext context, ClusterManager manager) {
         this.context = context;
         this.manager = manager;
 
@@ -58,7 +58,7 @@ public class MultipleAttributeAddition extends AbstractTask implements Rank {
     @Override
     public void run(TaskMonitor taskMonitor) {
         taskMonitor.setProgress(0.0);
-        taskMonitor.setTitle("Multiple Node Edge Additive ranking of clusters");
+        taskMonitor.setTitle("Multiple Node Edge Multiplum ranking of clusters");
         taskMonitor.showMessage(TaskMonitor.Level.INFO, "Fetching clusters...");
         taskMonitor.setProgress(0.1);
         clusters = ClusterUtils.fetchClusters(network);
@@ -131,18 +131,6 @@ public class MultipleAttributeAddition extends AbstractTask implements Rank {
         return clusters;
     }
 
-    private void setRankScore(String attribute, CyRow row, NodeCluster cluster) {
-        try {
-            cluster.setRankScore(cluster.getRankScore() + row.get(attribute, Double.class, 0.0));
-        } catch (ClassCastException cce) {
-            try {
-                cluster.setRankScore(cluster.getRankScore() + row.get(attribute, Integer.class, 0));
-            } catch (Exception e) { // Not a number type!
-                e.printStackTrace();
-            }
-        }
-    }
-
     private List<NodeCluster> setEdgeScoresInCluster() {
         List<NodeCluster> clusters = new ArrayList<>(this.clusters);
         List<CyEdge> edges = network.getEdgeList();
@@ -165,7 +153,7 @@ public class MultipleAttributeAddition extends AbstractTask implements Rank {
                     if (clusterNumber == sourceClusterNumber && (clusterNumber < sourceHighestClusterNumber || sourceHighestClusterNumber == -1)) {
                         setRankScore(edgeAttr, edgeRow, cluster);
                         sourceHighestClusterNumber = clusterNumber;
-                    } else if (clusterNumber == targetClusterNumber && (clusterNumber < targetHighestClusterNumber || targetHighestClusterNumber == -1)) {
+                    } else if (clusterNumber == targetClusterNumber && (clusterNumber < targetHighestClusterNumber || sourceHighestClusterNumber == -1)) {
                         setRankScore(edgeAttr, edgeRow, cluster);
                         targetHighestClusterNumber = clusterNumber;
                     }
@@ -174,6 +162,21 @@ public class MultipleAttributeAddition extends AbstractTask implements Rank {
         }
 
         return clusters;
+    }
+
+    private void setRankScore(String edgeAttr, CyRow source, NodeCluster cluster) {
+        try {
+            if (cluster.getRankScore() == 0.0) {
+                cluster.setRankScore(1.0);
+            }
+            cluster.setRankScore(cluster.getRankScore() * (source.get(edgeAttr, Double.class, 0.0) + 1.0)); // assumes values between 0.0 and 1.0
+        } catch (ClassCastException cce) { //
+            try {
+                cluster.setRankScore(cluster.getRankScore() * (source.get(edgeAttr, Integer.class, 0) + 1));
+            } catch (Exception e) { // Not a number type!
+                e.printStackTrace();
+            }
+        }
     }
 
     public static boolean isReady(CyNetwork network, ClusterManager manager) {
